Create CodeChecker
This code will detect any unbalanced code you put into a text file. It does this using a stack, and matching the top and bottom of the stack after it reads every line. If the stack is unbalanced at the end it will tell you that it is not balanced, and on what line the error is. ignores comments and block comments using a regex to removed all commented lines after it has already read the file.